Servidor/requirements.txt
spyne
lxml
flask
jsonschema
grpcio
grpcio-tools
fastapi
uvicorn
strawberry-graphql
pymongo